+++ GotLcg/GotLcg.Database.SqlServer/Data/BootstrapData/DATA_ChallengeIcon.sql
@@ -2,7 +2,7 @@
 
 DECLARE @challengeIconData AS TABLE
 (
-	[Id] INT NOT NULL,
+	[Id] TINYINT NOT NULL,
 	[Name] NVARCHAR(10) NOT NULL
 );
